requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-debug-toolbar==3.4.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docker==5.0.3
flower==1.0.0
generics==3.4.0
humanize==4.1.0
idna==3.3
iniconfig==1.1.1
kombu==5.2.4
packaging==21.3
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-postgres==0.7.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.1
